Update(Vue): Added reactive inter-panel features
Added the `vuex` code allowing `DiffPanel` to get data from the store that gets updated as soon as
the `v-textarea`s in the "Current" and "Old" `Panel`s are updated (like a inter-component
`v-model`).

BREAKING CHANGE: The `Panel` and `DiffPanel` reacts a bit differently and need one more attribute
`storeField` to work.